requirements.txt
Pillow>=9.0.0
httpx>=0.23.0
numpy>=1.22.0
